redo commits to avoid issues with wrongly commited file
